Added kolor and LuckPerms adapter.

+++ src/main/java/dev/piotrulla/crafthype/titles/bridge/MoneyResolver.java
@@ -4,10 +4,10 @@ import org.bukkit.OfflinePlayer;
 
 public interface MoneyResolver {
 
-    boolean has(OfflinePlayer player, double amount);
+    boolean has(OfflinePlayer player, int amount);
 
-    void deposit(OfflinePlayer player, double amount);
+    void deposit(OfflinePlayer player, int amount);
 
-    void withdrawl(OfflinePlayer player, double amount);
+    void withdrawl(OfflinePlayer player, int amount);
 
 }
